Update design and changes in presentation 1
Updated design that separate the title of the page and content of the page. Add in photo to the title box.  Remove the border of the body, landscaping the content. Add the user profile dropdown navigation in the navigation bar. Displaying the user info and the current centre info. Change the centreName, testName,  address pattern to contain at least one letter and special characters, digits enabled.